ML/agents/langchain/reAct/requirements.txt
python-dotenv
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0

